Fixed LUT mapping issues on some drivers
Due to precision issues, the LUT coords were off by a tiny margin, which led to dithering patterns being visible for values equal to or above 1 and equal to or below 0. Also instead of modifying the entire color range my simply multiplying, I changed it to a replacement of the saturate function, that takes the small epsilon into account, which should also be faster.